#8.14 Bugfix

+++ src/controllers/videoController.js
@@ -47,7 +47,11 @@ export const getEdit = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id);
     if (!video) {
-        return res.status(404).ender("404", { pageTitle: "Video not found" });
+        return res.status(404).render("404", { pageTitle: "Video not found" });
+    }
+    // type도 비교하기 때문에
+    if (String(video.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -59,6 +63,9 @@ export const postEdit = async (req, res) => {
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found" });
     }
+    if (String(video.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndUpdate(id, {
         title,
         description,
@@ -69,6 +76,13 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found" });
+    }
+    if (String(video.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
